# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Documentation",
        "description": "API documentation for your Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "name",
                                    "price",
                                    "quantity",
                                    "minimum_quantity",
                                    "visibility"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "description": "ID of the supplier",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Name of the product",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Smartphone X"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "string",
                                        "example": "Latest model with 128GB storage",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "description": "ID of the category",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number",
                                        "format": "float",
                                        "example": 599.99
                                    },
                                    "quantity": {
                                        "description": "Stock quantity",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "minimum_quantity": {
                                        "description": "Minimum stock quantity",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "visibility": {
                                        "description": "Visibility of the product",
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ],
                                        "example": "public"
                                    },
                                    "picture": {
                                        "description": "Product image (JPEG, PNG, JPG, max 2MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "supplier_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Smartphone X"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Latest model with 128GB storage",
                                                    "nullable": true
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1,
                                                    "nullable": true
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 599.99
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "minimum_quantity": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "visibility": {
                                                    "type": "string",
                                                    "example": "public"
                                                },
                                                "picture": {
                                                    "type": "string",
                                                    "example": "product_pictures/image1.jpg",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-14T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-14T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "supplier_id": "The supplier_id field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}